Refactoring the spinner
The original spinner had a bunch of implementation details
leaking all over the place - in particular, the creation of
the thread, and the channels to communicate with it. Taking
a page from the Erlangs, there is now a Spinner struct that
has an implementation of the client side start/stop calls,
and the internal spin call on the spinner side.

This makes it possible to

```rust
use utils::say::Spinner;

let spinner = Spinner::start();
long_function();
spinner.stop();
```

Which is heaps nicer. Also uses thread.join() to stop, since
we don't actually need to send signals back and forth.