basic delivery api <method> <path> working
Normalized http client

- get and post methods now return result objects.
- added helper for extracting a pretty json from a result
- error for unsupported HTTP method for cli api helper
- support data body sent for post requests

Enhance the error handling

Removed the raw panics and convert them to delivery errors. For
functions that are intended to complete an operation via Delivery's HTTP
API, there are two classes of errors: low level HTTP error and
unexpected HTTP response. Not totally happy with the approach here, but
it's a start.