Add stub for delivery token command
This patch adds a stub `delivery token` command. Eventually, this
command will prompt for a password, hit the API to obtain a token, and
write it to ~/.delivery/api-tokens.

At present, the command prompts for a password and records a dummy token
in ~/.delivery/api-tokens keyed off of the current server, enterprise,
and user. The key fields are read from cli config and can be set via
options on the command line.

Tokens are stored in a simple flat file format that looks like:
```
172.31.6.130,Chef,alice|exampletoken
172.31.6.130,Chef,seth|beef
172.31.6.130,acme,seth|cafe
172.31.6.98,Chef,seth|123
```

You want a single token file so that you can re-use a token across
different project checkouts that correspond to the same server,
enterprise, and user.

I originally explored using TOML for this, but ran into issues with
toml-rs [not properly round-tripping quoted keys][1]. While I think that
issue has been fixed, the flat format is probably less code. It assumes
that servers, enterprises, users, and tokens do not contain `|` nor `,`.

[1]: https://github.com/alexcrichton/toml-rs/issues/53

A small change was made to FromError for IoError to include detail about
the error message.