Open the change URL upon review
Add `utils::open::item` helper to open a URL using the system's default
browser.

When you run `delivery review`, we now try to open the change URL in a
browser. Tested so far only on OS X.

You can disable the open behavior using the flag --no-open. This is
especially useful if scripting with the delivery command line tool.