Return structured data from git_push_review
Instead of printing output from within git_push_review, we now return
structured data in the form of a ReviewResult struct. This contains the
parsed push results as a vector of PushResult structs, a vector of
String messages returned by the git server (those lines prefixed with
`remote: `) and some special handling for the remote messages to extract
a line that looks like a URL (current approach uses
`.starts_with("http")`).

This gives us some flexibility for prorammatically handling the results
of a push for review.

Here, we've reduced the output a bit. The git push results are not very
informative in the case of creating a review when talking to a Delivery
server since we will always be creating a new (temporary) branch. So the
output of that message is removed in favor of the remote message sent
from the Delivery server which will indicate whether a new patchset was
created or not.

Here's an example of the output we see for `delivery review` with this
change:

```
Chef Delivery
Loading configuration from /Users/seth/oc/code/opscode/delivery-cli
Review for change sf/git-cleanup targeted for pipeline master
Created new patchset
https://delivery.shd.chef.co/e/Chef/#/organizations/Chef_Delivery/projects/delivery-cli/changes/9a461f90-05d7-461f-9908-286bccfb196a
```