Add openssl to omnibus-delivery-cli
With the addition of hyper, the delivery cli now links to openssl which
needs to be included as an omnibus package dependency.

Also included is an update to latest omnibus and omnibus-software and a
few tweaks to the included omnibus.rb config to enable the S3 cache and
build in a local directory.