Allows cookbooks to be pushed to artifactory
Adds an artifactory resource and provider. Currently written in Chef 11
style, due to constraints of the existing build nodes. When we have new
build ndoes in place, we need to do two more things:

1) Add the credentials for artifactory to the builder on the chef run
list that sets the builders up

2) Update the code in publish to actually use them.

In the meantime, you can publish a build manually by uncommenting the
resources, putting the password in place, and running the publish phase
manually.