Wired up token support
The `delivery token` now hits the Delivery API and requests a token with
the password entered on the command line. If authentication is
successfull and a token is returned, it is written to the token store in
~/.delivery/api-tokens keyed by server, enterprise, and user.

Calls to `delivery api` will read the token to make requests. For
example:

```
delivery api get orgs
```