Add project import to the init command.
Example usage:

~/sample % delivery init --user=jra --server=192.168.33.66 --ent=Chef --org=Chef_Delivery --project=sample

If the 'sample' directory is not a git repo init will print an error and exit,
advising the user to 'git init' the directory.

If the 'sample' directory is a git repo but there are no commits, this will:

- Load configuration if it exists, or create it
- Run 'git init' on the 'sample' directory if it is not a repo
- Create the 'sample' project on the named server if it doesn't exist
- Attempt to create the master pipeline

Once commits are added to the local repo, or if they exist previously, re-running the command will:

- If the project's repo is empty, push any local commits to it
- If the project's repo has content, error out
- Create the master pipeline

Project import is integrated with the token store, so a 'delivery token ...' should be run before use.

Fixed a bunch of various compile-time warnings too.