Add behavioral tests using Cucumber
This adds a suite of Cucumber [1] tests for the Delivery CLI, using the
Aruba CLI testing library [2]. There are several things going on in this
patch, which bear mentioning.

Cucumber Test Features
======================

In order to test the CLI's interactions with the Delivery REST API, I
borrowed an approach used by Hub [3], a CLI tool that augments the `git`
command for interacting with Github. Like them, we now have a "fake API
server" that we can use in any test, whose behavior can be easily
modified to test a variety of scenarios. In order to easily do this,
however, I needed to make modifications to the CLI's Rust code as well,
which are detailed below

Another feature borrowed from Hub is the use of "fake binaries" for both
`git` and `open`. These fakes get added at the beginning of the tests'
`PATH` and will get used instead of the system's binaries. We can then
test behavior that would otherwise require a real git server (or web
browser, in the case of `open`). To ensure the proper things happen, we
log the commands that these fake binaries run and can check in our tests
to verify that the correct commands were executed.

We have several custom Cucumber steps for our scenarios (see
`features/support/steps.rb`). We will add more in the future, but these
are good for now. Many of these were either borrowed or based on steps
from Hub, as well.

Changes to the CLI
==================

We currently allow the port on which the Delivery Server listens for
HTTPS traffic to be configured. However, we hard-coded the standard
HTTPS port `443` into the CLI, which means not only would the CLI be
unusable for any customers that had made changes to their server
configuration, but tests involving any API interactions would need to be
run as root, in order for the fake server to use port `443`. As a
result, I modified the CLI to obey a user-specified API port (either in
a `.delivery/cli.toml` file, or passed as an argument), defaulting to
`443`.

While I was testing, I also realized we had a similar situation with the
Git port, in that we allowed the user to specify a configuration for the
CLI, which we actually ignored completely. That is fixed now, also.

In order to minimize the scope of these changes, I operated mainly at
the edges, at level of the CLI task implementations, rather than deeper
down in the code. All places where we originally accepted a
`server` (i.e. hostname) parameter would also accept a
`hostname:port`-style string as well (and indeed, some of the unit tests
showed this behavior already). As a result, we now create such
`hostname:port` strings based on configured values and pass this in
where we used to just pass in `hostname`.

Other
=====

The Cucumber tests will run in the unit phase of the Verify and Build
stages.

[1]: https://github.com/cucumber/cucumber
[2]: https://github.com/cucumber/aruba
[3]: https://github.com/github/hub