Add --version to delivery command line tool
The version is an iso8601 timestamp and the Git SHA from the build. On a
second line, version information from rustc is included.

The timestamp + git SHA version means we don't need to tag the repo nor
edit a file to modify the version on changes. The downside is that the
version does not provide semver information.

The git SHA and rustc information is compiled into the binary at build
time via environment variables. If these are unset, dummy values are
inserted (a SHA of 0000 and a message "rustc UNKNOWN"). This is to allow
a simple `cargo build` to succeed without going through a Makefile or
otherwise ensuring the environment variables are set.

Example output:

```
delivery --version
delivery 2015-06-07T00:34:18Z (fac367b)
rustc 1.2.0-nightly (99c2f779d 2015-05-29) (built 2015-05-28)
```