Fix missing port in git URL for cloning repos in jobs
Refactored the code somewhat as this particular error was pretty easy to
make with the previous structure. Now the function computing Git SSH URLs
takes a `Config` object as argument.
It seems to me, on a somewhat related note, that we should try and pass
`Config` objects around more, instead of having so many functions taking
scoping names as arguments.

Updated unit tests on that, and added a new Cucumber test on `delivery job`.
More Cucumber tests on that subcommand should be added as we go.

Also beefed up Cucumber tests some, notably by:
 * making more git commands effective (only `push` is not exectued now);
   we now actually clone repos from a local remote
 * making it possible to add ERB syntax in `... should be run` tests
   (mentioning that pattern here more as a thing to keep in mind if we need
   it again some place else)