Make unit and integration tests pass on Windows
This patch adds Windows-specific compatibility functions to allow
delivery-cli to pass tests on Windows.

The `utils` module now has two platform-specific implementation modules,
`unix` and `windows`, that are conditionally included/re-exported.

The `windows` module shells out to powershell to implement
copy_recursive and remove_recursive.

For Windows, the functionality for dropping privilege and changing
ownership and file mode are all no-ops for now.