PathFetcher "fix": clean local/src/delivery-cli
The omnibus PathFetcher currently relies on Ruby's filelib::cp but seems
to not clean the destination prior to the copy. And so the copy fails if
there are files that have read-only access mode (not also the danger of
extra files being left behind).

For now, we'll just cleanup this directory prior to starting the omnibus
build.