Teach make rustup to only rustup if needed
The `rustup.sh` script calls out to sudo if needed so we don't need to
prefix the rustup.sh command with sudo. We unconditionally add `--yes`
so there is no further user interaction (other than password). The
operation of what is happening via the makefile is clear enough that
users of the makefile shouldn't be surprised and the sudo password
request acts as a check anyhow.

The `rustup_builder` target is now identical to the `rustup` target and
can be removed once callers of it have been updated.