Use custom openssl and refactor build
The omnibus build for delivery-cli no longer bundles openssl. Instead,
openssl is installed on the build node and statically linked.

The top-level Makefile now accounts for the custom openssl
requirement. However, rust installation via rustup and the
installation of the openssl build-time dependency have moved out of
the Makefile and into the delivery_rust build cookbook.

The `delivery_rust::default` recipe can be used to bootstrap a system
for building delivery-cli. It will also serve to upgrade rust
appropriately for building delivery-cli within Delivery. To prepare a
system to build delivery-cli:

    cd cookbooks/delivery_rust
    berks vendor cookbooks
    sudo chef-client -z -o delivery_rust::default

On OS X, omit the `sudo`. This has been tested on OS X, Ubuntu 14.4,
and Centos 6.5.

The requirements for prepping the node to build a delivery-cli omnibus
package have also been simplified and only the required recipes from
the omnibus cookbook are included now via a new `_omnibus` helper
recipe.