Make postrm removal from /usr/bin conditional
On Redhat based systems, the postrm script for the old version is run
after the new version is installed upon package upgrade. Without
protection in the postrm script, we end up removing the
`/usr/bin/delivery` symlink on upgrade.

With this change, the removal will only occur on non-redhat OR on Redhat
where the action is uninstall (first arg to script == 0).