Fix config handling to use defaults
The default values for config items were being unconditionally
overwritten even if no value was provided in the parsed TOML data. With
this patch, the value from the TOML data is only set if there is
`Some(_)`. There is still a problem with the setup in that a non-string
value will map to `None` (and cause a default value, if present, to be
used silently).

Also fixing a bug that `pipeline` was not settable.