Add option to skip the default recipe
The `delivery job` subcommand now accepts `--skip-default` to skip
running the default recipe even if run by root.

Especially for local workflow, we want the option to skip the default
recipe for a faster cycle time when we are confident that the default
recipe has already been applied.

Note that this also helps us with the hack of running *only* the default
recipe via `delivery job verify default --skip-default` -- this is a bit
odd looking, but will prevent a double default run if run by root and
the right thing will happen for non-root users (where the skip-default
flag is not needed).