Allow multiple phases in single chef-client run
You can run multiple phases (intended only for local workflow) by
specifying a quoted argument to the job subcommand. For example:

```
delivery job verify "unit lint syntax"
```

This will launch a single `chef-client -z` run with all three phases in
the run list.

This patch also updates the job setup to ensure that the nodes/ and
cookbooks/ directories exist in the workspace. When skipping the default
recipe, these directories may not have been created and the chown call
that is part of the drop privileges dance will fail.