Add APIClient::from_config
Rather than extracting the values we need and having large arity
functions, let's start providing APIs that can consume the Config struct
and extract what is needed. This patch represents one piece of such a
refactoring.

We'd like to add a config option for `api_protocol` so that we can run
the cucumber tests on HTTP. Right now, HTTPS is hard-coded. Switching
things to pass the config struct will make the addition of api_protocol
cleaner as we won't have to add an additional protocol argument to the
call flow.