Add a token_file config option
You can now customize where delivery will look for the file containing
api tokens. The default remains as `~/.delivery/api-tokens`.

The motivation for being able to set this was to be able to test without
having to set an environment variable to relocate HOME in order to gain
test isolation.