Prompt for token when needed
As a user I would like to be quickly prompted for a token when I need it
instead of failing and asking me to run `delivery token`. That way I will
not be worry about when do I need token or not.

If we are using `delivery api`, `delivery init` or any other command that
requires a `token` now it will quickly prompt you to request one and store
it for further commands.