Make password prompt optional
Some times we dont want everything to ve interactive. This commit is giving us
the ability to specify a `non_interactive` option into the `cli.toml` file
so we can disable the password prompting that requires human interaction.

Added also some functions to be able to use `bool` options.

Added test for code validation and cucumber tests as well.