Delivery job locally should not change `HOME` dir
As a user that loves to test locally using the `delivery-cli` I would
like to consume my local `gems` and extra configuration I have on my
local workstation, so I don't have to install the dependencies on the
workspace on my own.

**Problem Description**

When executing `delivery job verify "unit syntax lint" --local` the cli
changes the `HOME` directory to `workspace/cache` causing me to lose all
the `gems` that I have in my `~/.chefdk`. Additionally as we do not run
the `default.rb` recipe that installs the dependencies of my
cookbook/project then I enter into a state where I can't run it without
manually installing the deps into the local `workspace/cache`.

**Proposal Solution**

Do not change the `HOME` directory if we pass `--local` as an argument.