Revamp command line parsing and help system
This patch replaces docopt with [clap][]. The clap crate provides better
command line parsing functionality for command line tools that have
extensive sub-commands such as delivery-cli. It allows for a mix of
docopt-style help string parsing as well as explicit builder pattern for
building the parser. Out of the box we get more readable help output as
well as sub-command specific help.

As a next step, we could leverage clap's `after_help` and the `include!`
macro to have subcommand-specific help overview which we can maintain in
a separate file.

[clap]: https://github.com/kbknapp/clap-rs