Add delivery checkout feature test
The checkout.feature test validates that delivery does the right git
invocations to do the checkout subcommand.

As part of this change, we simplified fakebin/git a bit. It was a bit of
a dance to get here and it still feels wrong. It seems that we need a
way of arranging for a feature-specific fakebin to handle mocking git
with the appropriate side-effects.