Workspace Path not hardcoded
Let me tell you a story. Currently we have the ability to customize the
workspace on the build nodes, this by configuring an attribute into the
`delivery_build` cookbook. That is cool. The problem is that in the
`delivery-cli` we do not have that same ability because the workspace
path is hardcoded. (See
https://github.com/chef/delivery-cli/blob/9e387f7415fbd9ac0d5f97c8ad523efe244b8722/src/main.rs#L605)

This commit is changing that behavior. Now the `delivery-cli` will not
have the workspace hardcoded, instead what we are going to do is take
advantage of the `delivery-cmd` script that is in change of trigger all
the `delivery jobs` coming from Delivery. This script sets up the `HOME`
directory to the actual `workspace` (that is customizable as we said)
and the `delivery-cli` will consume that and take it as the
`workspace` as well. Great!

But we are not done yet. As we know, the `delivery-cli` also exposes
some Workspace Attributes for the end-users so that they can customize
their `build-cookbooks`. If we think about it, the `workspace` path is
very useful as well since there are some files that the build-cookbooks
might use. Things like the `.chef/knife.rb` for the Chef Server
interaction, or the `.bin/git_ssh` that is a Git SSH Wrapper. (among
others)

This commit is exposing a new attribute called `workspace_path` that any
build cookbook can use by accessing the attribute:
`node['delivery']['workspace_path']`

Added documentation and some `debug!` logs for debug purposes.