Added mocking tests for Private Supermarket
This tests are designed to test the Private Supermarket aproach, instead
this test is ponting to the same public supermarket since it will be
always up and running.

We should put this kind of tests in Acceptance where we might be able to
use a Private Supermarket from Union. :smile:

Updated README Docs