Disable Omnibus’s git cache to prevent versioning issues
We pass the version info (which is returned by `delivery --version`) 
into the Rust compile using the `DELIV_CLI_VERSION` and `DELIV_CLI_GIT_SHA` 
environment variables. There exists the possibility that a re-build of
the `build/publish` stage will cause the output of the `delivery-cli` 
build to be fully restored from the git cache and thus have the version
number of the previous build. This doesn’t sound like a big deal if we 
were building for a single platform but with our larger build matrix this
means certain platforms could return a different version number when 
`delivery --version` is run. 

We’ll work around this issue by disabling the git cache feature, as a full 
build of the `delivery-cli` project only takes ~ 3 minutes.