Verify SCP configuration in the Delivery Server
Added the SCP verification to fail fast in the case that the SCM is not
configured in advance in the Delivery Server. By default the Github
integration will say always `true` but this change is ready to pick the
configuration from the Server side when we apply it. On the other hand,
the Bitbucket configuration is working and it will fail if the user
hasn't configured on the Server side.

Also pointed Berksfile to supermarket in the build-cookbook and added
extra debug outputs.