Verify token when loading APIClient config
Now the `APIClient::from_config` not only will get the config from
the tstore and pass it through but also verify that the token hasn't
expired by hitting the endpoint get("orgs") - If the token has expired
then we will notify the user and request a new one. To avoid this use
the `--no-interactive` flag to throw and `Err()`.