Handle project creation when it already exists
Before we just verify if the project already exists in Delivery and we
skip the rest of the server interaction, that is pushing the local
content to Delivery/Git and create tht pipeline.

Now we are verifying each thing. If the project exists already, we now
proceed to verify that we have content on the remote repo and that the
pipeline exists, if not we will create it.