Remove build processes since delivery-cli now ships in the ChefDK.
Since we now only ship delivery-cli in the ChefDK, we should not build and package it seperately. This commit removes the omnibus part of the project, as well as removes omnibus building and artifactory publishing from the build cookbook.

It also moves all the github publishing to acceptance/functional, which is the general spot it shows up in our other projects. This way, when you hit "deliver", the code will show up on github so it can be pulled into the ChefDK.