Use bundled `chef generate build-cookbook` instead of pcb and rm config generation.
We are deprecating pcb in favor of `chef generate cookbook`.

Also, it doesn't make sense to have config generation in multiple places, so also removing generation code from DeliveryConfig, as it will be generated by default in `chef generate build-cookbook`. If people are using custom generators, they should be either passing in a custom config or their generator should be generating one.