Cleanup logic around config generation in init and tests.
Since `chef generate build-cookbook` gives you the `.delivery/config.json`, we should only have that config generator in one place, so it has been removed from delivery-cli, which resulted in a bit of cleanup around the logic of config generation.

Also, removed a bunch of mocked and beefed up the tests.