Avoid caching values on workspaces
How this works? Delivery triggers the jobs to the build-nodes, every job
has a workspace per phase with three important directories:

* repo  - The cloned repository
* cache - A simple cache directory to use.
* chef  - The chef configuration, build-cookbook, cookbooks, dna.json,
nodes and more.

The main problem itself is that these directories doesn't get cleaned,
that means that there is always stale data that get in your way.
Essentially because we are using the delivery-cli to trigger (under the
hoods) a chef-zero run that loads all these directories/data and merges
the new data instead of overwritten it.

Solution
========

Clean the workspace::chef/nodes directory

We have to clean the `nodes/` directory since `chef-zero` will merge the
old attributes with the new ones of the coming change. By deleting this
directory and creating it clean we will ensure that the values we lay down
are fresh and clean.

Support Ticket: https://getchef.zendesk.com/agent/#/tickets/9925
Closes https://github.com/chef/delivery-cli/issues/15