[ET-248] Fix 'delivery token' for SAML authentication
Before, Delivery offered an endpoint for looking up if a specific user
was SAML-enabled or not. This was used when the user invoked `delivery
token` to choose the right authentication flow.

Recently, this endpoint was replaced by a enterprise-wide "is SAML
enabled?" endpoint.

This change changes the behaviour of `delivery token` to: lookup if
Delivery is configured for SAML, and if so, uses the SAML authentication
flow.

Since this is an all-or-nothing affair wrt. users, it also introduces
both a CLI flag for `delivery token`, and an option for
`.delivery/cli.toml`, to override it: `saml=true/false`.