Mark cucumber test broken for delivery init --for
This test is currently broken

When we run `delivery init` for a different pipeline that is not master
we would expect the repository to be already inside the pipeline-branch
and also to have at least one commit. (as usual)

Problem: ChefDK internally is taking care of the repository manipulation,
that is, adding the build-cookbook, delivery config.json and more. When it
is done with the modification it also commit the changes to master and
push the branch to the Delivery Server. As this code is currently hard
codded to be always pointing to `master`
=> https://github.com/chef/chef-dk/blob/master/lib/chef-dk/skeletons/code_generator/recipes/build_cookbook.rb#L146

We have two options to fix this:
1) Pass an option to chefdk generate with the branch-pipeline name
2) Remove the git commands from chefdk and let the delivery-cli handle them

TODO: Uncomment this like once we fix this problem