Create DELIVERY.md in feature branch on init
`delivery init` must create a feature branch and trigger a review, or
else users will not know if subsequent changes fail because the changes
are bad or because the pipeline never worked in the first place.
`chef generate` is now responsible for creating the delivery
configuration in a feature branch AND merging it; this is the desired
behavior for users who do not have a delivery server. Therefore, we add
a documentation-only change and trigger a review of that in order to
exercise the pipeline when initializing a project.