Fix token verification on every request
This change is fixing and old behavior that was broken.

Fixed the token verification on every single request that we use
internally. Now when users use the CLI and their token is expired,
instead of throwing an Err() now we will regenerate the token
by requesting a new one.

TODO: Upgrade Aruba to allow broken cucumber test to run.