Generate the delivery config before the build_cookbook
If the user provides the config, we should lay it down first and then
generate the build_cookbook. The reason why this is important is because
the generator might not have the steps to lay it down and this causes
the delivery init command to fail saying you dont have a config file.

This will allow users to run:
```
delivery init --generator https://github.com/afiune/enterprise_chefdk_generator -c ~/config.json
```