Generate build-cookbook with build-cookbook command
Now that chefdk has a command to generate cookbooks, we should stick
with the command and use it to generate the custom build-cookbook
instead of trying to use the old `chef generate cookbook`.

This will give us a big advantage that is, customers can use the same
generator to generate cookbooks and build-cookbook without having
to decouple.