Honor the custom config when it is passed on init
Before when we run the command:
```
$ delivery init -c /my/custom/config.json --generator /path/to/custom_generator
```

The delivery-cli failed saying that the build-cookbook didnt create the
config.json. That seems to be the wrong approach since we also passed
a custom config.

This commit is honoring the custom config, that means that from now on
we will use the custom generator and then the custom config and after
both have happened we will verify that we have a `config.json` inside
the `.delivery` directory.