Added gem deps to run chefdk commands in cucumber
If we need to run some chefdk commands that has gems dependencies we
have to add them to the `Gemfile` so that bundle has it and therefore we
can use them within cucumber.

This commit is specifically adding the `rspec` gem dependencies to run
`delivery local unit` that inside is running `chefspec` tests.