Run cli-tests using lint phase instead of unit
One of the cli-test is using the unit phase to verify how the jobs are
being rendered and dispatched on the build-nodes, the problem is that we
are having some failures with berkshelf where it blows up with the
error:
```
E, [2016-10-05T15:39:19.983072 #29299] ERROR -- : Actor crashed!
```

To stablelize the pipeline I propose we use lint instead since it is
more simple and it just uses rubocop which is more secure.

Signed-off-by: Salim Afiune <afiune@chef.io>