[COOL-598] Ensure Git can be found on Windows ChefDK installs
Since only `git.bat` is found in the `PATH` on Windows ChefDK installs:

https://github.com/chef/omnibus-software/blob/master/config/software/git-windows.rb#L50-L52

`delivery-cli` commands that shell out to `git` fail because Git can
not be found when Rust tries to find the executable to be spawned. This
is because Rust's Windows spawn command, when provided with a file path
without an extension, will only search for `.exe` files when searching the path:

https://github.com/rust-lang/rust/issues/37380

https://github.com/rust-lang/rust/blob/master/src/libstd/sys/windows/process.rs#L141

https://github.com/rust-lang/rust/blob/master/src/libstd/sys/windows/env.rs#L18

This change adds a `find_command` that uses the `PATHEXE` env variable on
Windows to also search for other valid Git executables in addition to `.exe`.

http://environmentvariables.org/PathExt

This is the same approach Habitat took when they ran into this issue:

https://github.com/habitat-sh/habitat/pull/1290/files#diff-5fb0c7b1056276e7845dbf93cb430542R258