Fix bug where delivery remote can get out of date.
We initially set up the git remote on clone with settings populated by server, user, ent, org, and git port.

These settings might change if those get updated but those changes are never reflected in an updated remote.

For example, if your git port changes, then your delivery git remote needs to reflect that.

This change / fix will update the remote based on relevant cli.toml / passed in settings every time a command specific to a project is run.

This change / fix is needed as part of the fips work, which will be setting the delivery remote to a localhost:<specific_port>.