[COOL-640] Add option to load project.toml from URL
This commit adds an `-r` option to the `delivery local` command:
```
OPTIONS:
    -r, --remote-project-toml <remote-url>    URL for remote project.toml
```
This allows users to load and use a `project.toml` from a URL. Without the
flag `delivery local` will continue to try and use the local `project.toml`
on disk in `.delivery`. When a URL is provided, the remote `project.toml`
takes precedence over the local `project.toml`.