Add fips mode.
This adds fips and fips_port config values to cli.toml. Setting fips to true will update your remote to point at localhost:<default_fips_ports> where we expect stunnel to be listening to tunnel fips-encrypted git requets to automate (configuration currently needs to happen seperately, as this is a WIP feature). If you are not using the default port defined in this change for your stunnel accept, you can set the fips_port value.