Make local phases optional
This commit is making all the local phases optional. That means that we
can have partial project.toml files that now are able to be parsed and
executed.

If a user is running a local phase that is not defined, the cli will
fail with a nice message.

This will also help moving the remote_toml feature to an state where we
don't have to define the phases and instead we just need to point to the
remote project.toml file.

Signed-off-by: Salim Afiune <afiune@chef.io>