Implement Display for Phase enum
We want to display the phase name with lowercases instead of capital
letter.

Now when a user tries to execume `delivery local unit` when the phase is
not defined in the `project.toml` you will see a message like:
```
Chef Delivery
Unable to execute an empty phase.
Please verify that your project.toml has a unit phase configured as follows:

[local_phases]
unit = "insert script here"
```

Signed-off-by: Salim Afiune <afiune@chef.io>