Do not unwrap() Results, use try!() instead
This was originated as a Serv1 where all the jobs start failing with
```
Chef Delivery
thread 'main' panicked at 'called `Result::unwrap()` on an `Err` value: DeliveryError { kind: NoGitConfig, detail: None }', ../src/libcore/result.rs:788
note: Run with `RUST_BACKTRACE=1` for a backtrace.
```

The main problem is that we was unwrapping Results instead of passing
them through the end of the execution.

This commit is not changing any functioanlity, it is just modifying the
unwrap statements and implementing them with `try!()` instead.

Signed-off-by: Salim Afiune <afiune@chef.io>