FIX: Dont explode if you cant find the project_path
This is the actual fix for the sev1 we have yesterday, it was actually a
combination of things that at this point will be fixed.

Prob we might need to re visit the implementation and use case of the
trait `InitCommand` since it does not seem to be clear why we have it.
But we will let that for another change with more time.

Signed-off-by: Salim Afiune <afiune@chef.io>