Enable load_config to load DeliveryConfig V1 & V2
We have the DeliveryConfigV1 ready to be depricated, but as a good
practice for backwards compatibility we will still supported.

This commit enables `load_config` to be able to load Config V2, which is
the latest as this commit, but if it can't decode V2, it will try to
decode V1 and convert it to V2.

Signed-off-by: Salim Afiune <afiune@chef.io>