Verify build-cookbook path on project init
Currently the `chef generate build-cookbook` command doesn't allow
you to pass a custom path to generate a build-cookbook out side the
default location that is `.delivery/build_cookbook`. Thereof we need
to verify the path and if it is custom we will `Err()` with a helpful
message.

TODO: (IDEA#383) Be able to generate build-cookbooks on a custom location

Signed-off-by: Salim Afiune <afiune@chef.io>