Tests to verify regression when rendering DNA file
This test is verifying that, when a project has custom attributes inside
the `.delivery/config.json` they has to be available on every phase so
that they get used by the build_cookbook.

A clear example of this is in the `delivery-truck` build_cookbook:
=> https://github.com/chef-cookbooks/delivery-truck/blob/master/.delivery/config.json#L10-L22

This is necessary until we use a reserved work like `attributes` or somethig
else that will allow us to be prescriptive about the content of the file that
can be configurable, until then we need to use a very generic Json format.

Signed-off-by: Salim Afiune <afiune@chef.io>