Fixed regresion in DNA file; use generic Json
There are projects that have custom attributes inside the
`.delivery/config.json` that needs to be available on every
phase of the pipeline so that the build_cookbook can use them.

A clear example of this is in the `delivery-truck` build_cookbook:
=> https://github.com/chef-cookbooks/delivery-truck/blob/master/.delivery/config.json#L10-L22

For this reason we need to use a generic Json format until we
have a reserved word in the config that allow us to be prescriptive
about the content of the file that can be configurable.

TODO: Restrict the config.json format by selecting a reserved field
that allow users to inject attributes that will be used by the
build_cookbook.

An idea could be for example a field called `attributes`:
```
{
  "version": "2",
  "build_cookbook": {
    "name": "cerebro",
    "enterprise": "marvel",
    "organization": "x-men"
  },
  "job_dispatch": {
    "version": "v2",
  },
  "skip_phases": [],
  "dependencies": [],
  "attributes": {}             <--- Inject here any attribute you need
}

```

Signed-off-by: Salim Afiune <afiune@chef.io>