Verify delivery git remote and ask for permissions
If we are about to modify something on the current repository it would
be a great UX if we notify the user we are going to do this, and even
better, we should ask for permission to do so.

This commit includes:
* delivery git remote verification - Now we verify that the remote is
actually not up-to-date.
* request confirmation of git remote modification - If the remote needs to
be updated, now we will prompt the user to confirm that this is something
he wants to proceed with.
* an idea to write the `cli.toml` inside every repository that the user
clones, this way we could ensure that we always have the right config
within.

Example:
```
The 'delivery' remote doesn't match with the default configuration loaded from "/Users/salimafiune/chef/delivery/organizations/sandbox/.delivery/cli.toml".
	current: ssh://afiune@chef@delivery.example.com:8989/chef/sandbox/foo
	update:  ssh://afiune@chef@localhost:21334/chef/sandbox/foo

Would you like to update it? (y/n):
```

Signed-off-by: Salim Afiune <afiune@chef.io>