Do not prompt for remote update verification
After an internal conversation we have decided to avoid asking the user
to validate the modification of the remote since that could potentially
break the runners job dispatch since they can't answer the question. :)

NOTE: If we want to do that we could leverage the `--no-interactive`
flag. Just need to pass it to the `delivery-cmd` in Chef Autoamte.

Signed-off-by: Salim Afiune <afiune@chef.io>