Fix utils test & delete unused macro import
The `traverse_up_for_dot_delivery_found` test was pointing to the current_dir
that used to have a `.delivery` folder. Since we don't have it anymore we are
pointing it to a test_repo that has one.

Also deleted the unused macro imports from `main.rs` and `lib.rs`

Signed-off-by: Salim Afiune <afiune@chef.io>