Verify the remote always on project initialization
When we are initializing the project we must always verify that the
remote is up-to-date since we are actually configuring the project on
the Automate Server.

This commit is adding a new section to add/verify/update the delivery
git remote.

Also is deleting it from the `Command::setup` fun since we don't need it
twice! :smile:

Signed-off-by: Salim Afiune <afiune@chef.io>