Add a2-mode option to any command that takes --server
For consistiency we should have the a2-mode flag available when we
have a command that lets you override server, since it's principal
function is to modify the resource that formed using the server
argument.

I wrestled with forcing the saml-mode check in all cases, but in the
end decided that wasn't necessary because the saml mode option may not
be appropriate in all cases. It's worth reviewing that again for
sanity.

Signed-off-by: Mark Anderson <mark@chef.io>